+++ src/app/(graybg)/Layout.module.css
@@ -1,0 +1,20 @@
+.wrapper {
+  display: flex;
+  justify-content: center;
+  height: 100dvh;
+  background-color: var(--color-bg-secondary, #fafafa);
+}
+
+.container {
+  margin: 0 74px;
+  width: 100%;
+  /*min-height: 100dvh;*/
+  max-width: 1140px;
+}
+
+@media screen and (max-width: 768px) {
+  .container {
+    margin: 0 20px;
+    min-width: 280px;
+  }
+}
