+++ src/app/(graybg)/layout.tsx
@@ -1,0 +1,11 @@
+import styles from './Layout.module.css';
+
+const RestLayout = ({children}: {children: React.ReactNode}) => {
+  return (
+    <div className={styles.wrapper}>
+      <div className={styles.container}>{children}</div>
+    </div>
+  );
+};
+// ${styles[theme]}
+export default RestLayout;
