+++ src/app/(graybg)/users/_components/my/ProfileCard/ProfileCard.module.css
@@ -4,8 +4,7 @@
   align-items: flex-start;
   gap: var(--spacing-12, 12px);
   align-self: stretch;
-  border-radius: var(--border-radius-lg, 8px);
-  background: var(--color-bg-inverse-bolder, #262626);
+  z-index: var(--z-index-info-box-content);
 }
 
 .bg-image__wrapper {
