+++ src/app/(whitebg)/users/_components/steps/JoinCompleted.tsx
@@ -1,7 +1,7 @@
 import styles from './JoinCompleted.module.css';
 import Image from 'next/image';
 import GrayMinari from '@/assets/minari-gray-cut.svg';
-import ProfileCard from '@/app/users/_components/my/ProfileCard';
+import ProfileCard from '@/app/(graybg)/users/_components/my/ProfileCard';
 
 interface JoinCompletedProps {
   name: string;
